improve: O-Freeze v2.3.0 - detune knob is now true cents (2^(c/1200)), was linear 1+c/1200 (Stage O batch O1)
Item 48 (DSP) FIXED. PluginProcessor.cpp:582 mapped DETUNE as
playbackRate = 1 + r*(cents/1200); a cent is 2^(1/1200). Now
std::pow(2, r*cents/1200); the per-grain random spread r is untouched.
Minor bump: the sound of an existing session changes at the same knob value.

Follow-on FIXED, found by the measurement: Grain::fractionalPosition was a
float accumulator, so at a 1000 ms grain (44100 samples, ulp 2^-8) the
per-sample rate rounded to a multiple of 1/512 and every grain's pitch was
quantised to a 3.38 ct grid (1.69 ct at the default 400 ms) against a knob
that offers 0.1 ct. Now double.

Measurement (scratchpad offline render, not committed; processor at block
size 1, frozen 440 Hz sine, GRAIN_COUNT 2, DRIFT 0, LFO depth 0, one seed
both sides so every grain pairs; pitch by zero-crossings where one grain is
alone in the overlap-add; 44 grains per knob value, 209 in the 240 s run):
  knob  5:  before max +6.75 / -6.78 ct (only 0, +-3.38, +-6.75 occur)
            after  max +4.56 / -4.70 ct, continuous
  knob 25:  before +33.49 / -34.15  ->  after +22.79 / -23.51
  knob 50:  before +63.08 / -68.98  ->  after +45.59 / -47.01
            per-grain before/after ratio 1.29..1.63, median 1.45
            (theory: 1.41 at +50, 1.47 at -50)
  knob 50, 209 grains: before +69.59 / -72.51 -> after +49.50 / -49.96
  after the fix each grain's pitch at knob 25 is 5.00x its knob-5 pitch and
  at knob 50 10.00x (medians 5.000 / 9.999); the freeze-engage grain (rate
  1.0 by design) reads 0.000 ct both sides (negative control).

Copy: no English or French changed; tip.detune's "0 to 50 cents" /
"De 0 à 50 cents" was the claim and is now true. i18n.js header notes it.

Gates: check-ui-labels 0 moved before and after; check-i18n ALL PASS;
i18n-fr-lint 0 (--strict exit 0); tests/ui_tip_render_check.js 311 passed;
boot-all-uis 43/43 clean, 0 DEAD, O-Freeze late count 0 unchanged;
build-and-install under the Stage K mutex; auval aufx OFCR OuDv PASSED;
installed AU and VST3 Info.plist 2.3.0; header string "(v2.3.0)" present
in both installed binaries, the 2.2.1 header string absent.

+++ plugins/O-Freeze/Source/PluginProcessor.h
@@ -94,7 +94,12 @@ private:
         int position = 0;         // Base position in freeze buffer (set at activation, fixed)
         int jitterOffset = 0;     // Per-grain random position offset (decorrelates overlapping grains)
         float playbackRate = 1.0f;        // Pitch micro-detuning rate multiplier
-        float fractionalPosition = 0.0f;  // Fractional read offset from base position
+        // Fractional read offset from base position. DOUBLE, not float (v2.3.0):
+        // a float accumulator at 44100 samples (1000 ms grain) has an ulp of
+        // 2^-8, so adding a rate of 1.004 per sample rounds the rate itself to a
+        // multiple of 1/512 -- every grain's pitch snapped to a 3.38 ct grid,
+        // measured. The knob claims 0.1 ct.
+        double fractionalPosition = 0.0;
         bool active = false;      // Grain active flag
     };
